:truck: move ContextFilter to own module
